push from server

+++ database/migrations/2025_04_08_131427_create_crm_tasks_table.php
@@ -22,7 +22,7 @@ return new class extends Migration {
 
             $table->morphs('related_to');
 
-            $table->unsignedInteger('assigned_to');
+            $table->unsignedBigInteger('assigned_to');
             $table->foreign('assigned_to')
                 ->references('id')
                 ->on('users');
